feat: funcs from potential
Function-oriented approach.
