+++ src/galax/potential/__init__.pyi
@@ -11,13 +11,6 @@ __all__ = [
     # composite
     "AbstractCompositePotential",
     "CompositePotential",
-    # param
-    "ParametersAttribute",
-    "ParameterCallable",
-    "AbstractParameter",
-    "ConstantParameter",
-    "UserParameter",
-    "ParameterField",
     # builtin
     "BarPotential",
     "HernquistPotential",
@@ -40,6 +33,21 @@ __all__ = [
     "MilkyWayPotential",
     # frame
     "PotentialFrame",
+    # funcs
+    "potential_energy",
+    "gradient",
+    "laplacian",
+    "density",
+    "hessian",
+    "acceleration",
+    "tidal_tensor",
+    # param
+    "ParametersAttribute",
+    "ParameterCallable",
+    "AbstractParameter",
+    "ConstantParameter",
+    "UserParameter",
+    "ParameterField",
 ]
 
 from ._potential import io
@@ -66,6 +74,15 @@ from ._potential.builtin.nfw import (
 from ._potential.composite import AbstractCompositePotential, CompositePotential
 from ._potential.core import AbstractPotential
 from ._potential.frame import PotentialFrame
+from ._potential.funcs import (
+    acceleration,
+    density,
+    gradient,
+    hessian,
+    laplacian,
+    potential_energy,
+    tidal_tensor,
+)
 from ._potential.param import (
     AbstractParameter,
     ConstantParameter,
